luvit: add process.exitCode
It has happened a few times that the luvit event loop has exited before
we expected in our application due to bugs in luvit. It would be great
to be able to set the exitCode so that if we exit the loop before our
application expects it we can have a failing exit code.

Example usage:
    -- exit luvit with exit code 3 in the event of a pre-mature
    -- event loop exit
    process.exitCode = 3