lib: luvit: bind to new uv_ref/uv_unref signatures
uv_ref and uv_unref now work on uv_handle_t not uv_loop_t so replace all
calls to native.unref and native.ref with Stream/handle operations.

This is a WIP. It simply gets the luvit repl working again- tests do not
pass. Also, in order for this to work on OSX it needs a libuv PR

    https://github.com/joyent/libuv/pull/491/files