tests: test-tls-client-econnreset
Introduce a test for the tls client getting an error in the middle of
the stream. This found a couple of bugs which are fixed early in this
set of commits.

Aside: While writing this test I did something stupid.  I was trying to
pcall child.stdin.write with self being passed in being child.stdin.

I think I need to send a patch to libuv to have uv_write send back an
error instead of asserting so we can handle that more gracefully, easy
mistake to make.

Also increase logging around this test a bit.