uv: stop using closed and use close
in luvit we had some weird semantics around the difference between the
"closed" event and the "close" event. The closed event only came from
handles but then that trickled up into some of the other classes.

So, lets delete this closed state and only use close.

Close from here on out means the underlying fd that the Handle owns is
closed.