fix(deps/dns.lua): resolv.conf parser enhancements
This changes the parser so that it:
- ignores whitespaces at end of line
- treats lines starting with ';' as comments
- expects keywords to start at the beginning of the line.

From man resolv.conf:
Lines that contain a semicolon (;) or hash character (#) in the first column are treated as comments.
The keyword and value must appear on a single line, and the keyword (e.g., nameserver) must start the line.  The value follows the keyword, separated by white space.